requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
google-generativeai
requests
websockets
python-dotenv